rename package to follow ktlint rules
