+++ src/main/kotlin/cryptoservice/exception/GlobalExceptionHandler.kt
@@ -1,7 +1,7 @@
-package crypto_service.exception
+package cryptoservice.exception
 
-import crypto_service.exception.exceptions.SOPSDecryptionException
-import crypto_service.exception.exceptions.SopsEncryptionException
+import cryptoservice.exception.exceptions.SOPSDecryptionException
+import cryptoservice.exception.exceptions.SopsEncryptionException
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.web.ErrorResponse
